requirements.txt
numpy==1.23.4
pyopencl==2023.1
ffmpeg-python==0.2.0
